+++ Source/DlgSystemEditor/Private/DialogueSearch/DialogueSearchResult.cpp
@@ -79,7 +79,7 @@ TWeakObjectPtr<const UDlgDialogue> FDialogueSearchResult_DialogueNode::GetParent
 
 TSharedRef<SWidget>	FDialogueSearchResult_DialogueNode::CreateIcon() const
 {
-	const FSlateBrush* Brush = FDialogueStyle::Get()->GetBrush(FDialogueStyle::PROPERTY_DialogueClassIcon);
+	const FSlateBrush* Brush = FDialogueStyle::Get()->GetBrush(FDialogueStyle::PROPERTY_DlgDialogueClassIcon);
 
 	return SNew(SImage)
 			.Image(Brush)